move model inside camera frame
